move the mdns packages p2p/discovery
