+++ examples/chat-with-mdns/mdns.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
-	discovery "github.com/libp2p/go-libp2p/p2p/mdns_legacy"
+	mdns "github.com/libp2p/go-libp2p/p2p/discovery/mdns_legacy"
 )
 
 type discoveryNotifee struct {
@@ -21,7 +21,7 @@ func (n *discoveryNotifee) HandlePeerFound(pi peer.AddrInfo) {
 //Initialize the MDNS service
 func initMDNS(ctx context.Context, peerhost host.Host, rendezvous string) chan peer.AddrInfo {
 	// An hour might be a long long period in practical applications. But this is fine for us
-	ser, err := discovery.NewMdnsService(ctx, peerhost, time.Hour, rendezvous)
+	ser, err := mdns.NewMdnsService(ctx, peerhost, time.Hour, rendezvous)
 	if err != nil {
 		panic(err)
 	}
